refactor: BankClerk 삭제하고 Client에 기능 추가
