+++ BankManagerConsoleApp/BankManagerConsoleApp/Client.swift
@@ -17,7 +17,25 @@ class Client: Operation {
     }
     
     override func main() {
-        let clerk = BankClerk()
-        clerk.handleClientBusiness(of: self)
+        handleClientBusiness(of: self)
+    }
+    
+    func handleClientBusiness(of client: Client) {
+        print(ConsoleOutput.currentProcess(client, .start).message)
+
+        switch client.business {
+        case .deposit:
+            Thread.sleep(forTimeInterval: 0.7)
+        case .loan:
+            Thread.sleep(forTimeInterval: 0.3)
+            waitForLoanPermission(of: client)
+            Thread.sleep(forTimeInterval: 0.3)
+        }
+        
+        print(ConsoleOutput.currentProcess(client, .done).message)
+    }
+    
+    private func waitForLoanPermission(of client: Client) {
+        headQuarter.handleLoanQualificationQueue(of: client)
     }
 }
